Add value field to link + LineWidth + AutoColorLink
defaultLineWidth is the default link line width, it is multiplied by linkValueField to draw larger line for important links.
Warning: Due to limitations in the ANGLE layer, with the WebGL renderer on Windows platforms linewidth will always be 1 regardless of the set value. (three.js documentation)
linkAutoColorBy works identically to nodeAutoColorBy.